Move app/bin
