+++ .envrc
@@ -1,6 +1,6 @@
 use flake ./dev/nix/shell
 PATH_add node_modules/.bin
-suppress=$(expand_path ./suppress.cjs)
+suppress=$(expand_path ./dev/node/suppress.cjs)
 export NODE_OPTIONS_PRODUCTION="--unhandled-rejections strict"
 export NODE_OPTIONS="--require ${suppress} --experimental-json-modules --experimental-import-meta-resolve ${NODE_OPTIONS_PRODUCTION}"
 export DBMATE_MIGRATIONS_DIR=db/src/migrate
