+++ package.json
@@ -5,16 +5,16 @@
   },
   "type": "module",
   "scripts": {
-    "build": "node --experimental-loader ./loader.js app/bin/build.ts",
+    "build": "node --experimental-loader ./dev/node/loader.js dev/node/build.ts",
     "check-fmt": "prettier --ignore-path .gitignore --check '**/*.cjs' '**/*.js' '**/*.json' '**/*.ts' '**/*.html' '**/*.md' '**/*.yaml' '!pnpm-lock.yaml' && dprint check",
     "check-lint": "eslint --ignore-path .gitignore --max-warnings 0 --ext ts,js,cjs .",
     "check-tsc": "tsc",
     "db": "dbmate --url ${DATABASE_URL_ADMIN}",
-    "dev": "node --experimental-loader ./loader.js app/bin/dev.ts",
+    "dev": "node --experimental-loader ./dev/node/loader.js dev/node/local.ts",
     "fix-lint": "eslint --ignore-path .gitignore --fix --ext ts,js,cjs .",
     "fmt": "prettier --ignore-path .gitignore --write '**/*.cjs' '**/*.js' '**/*.json' '**/*.ts' '**/*.html' '**/*.md' '**/*.yaml' '!pnpm-lock.yaml' && dprint fmt",
     "start": "pnpm build && NODE_ENV=production NODE_OPTIONS=\"${NODE_OPTIONS_RELEASE}\" APP_PORT=\"${APP_DEV_PORT}\" node build/node/app/src/api/index.ts.mjs",
-    "test": "node --experimental-loader ./loader.js app/bin/test.ts",
+    "test": "node --experimental-loader ./dev/node/loader.js dev/node/test.ts",
     "watch": "tsc --watch"
   },
   "dependencies": {
